backend/requirementsss.txt
flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
google-generativeai==0.7.2
supabase==2.6.0
python-dotenv==1.0.1
